backend/requirements.txt
claude-agent-sdk~=0.1.39
supabase~=2.10.0
fastapi~=0.115.0
uvicorn[standard]~=0.32.0
python-dotenv~=1.0.0
pydantic~=2.10.0
